# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MathLab Pro API
  version: 0.1.0
  description: >
    Contract-first API for MathLab Pro. The current TypeScript API and the
    Java 21/Spring Boot backend must satisfy this contract.
servers:
  - url: http://localhost:3000
    description: Current local API server
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/cache/status:
    get:
      tags: [Operations]
      summary: Return calculation cache status
      responses:
        "200":
          description: Cache status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/audit-logs:
    get:
      tags: [Operations]
      summary: List audit logs for the current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 250
            default: 100
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            additionalProperties: true
...
    User:
      type: object
      required: [id, username, email, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 40
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    CacheStatus:
      type: object
      required: [provider, enabled, ttlSeconds]
      properties:
        provider:
          type: string
          enum: [memory, redis]
        enabled:
          type: boolean
        ttlSeconds:
          type: integer
        redisUrl:
          type: string
    AuditLog:
      type: object
      required: [id, action, resource, metadata, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        action:
          type: string
        resource:
          type: string
        resourceId:
          type: string
        metadata:
          type: object
          additionalProperties: true
        ipAddress:
          type: string
        userAgent:
          type: string
        createdAt:
          type: string
          format: date-time
...